xrRenderer: common types introduced
