+++ src/Layers/xrRenderPC_R1/stdafx.h
@@ -11,7 +11,7 @@
 #include <d3d9.h>
 #include <d3dx9.h>
 
-#include "Layers/xrRender/xrD3DDefs.h"
+#include "Layers/xrRenderDX9/CommonTypes.h"
 #endif
 #include "Layers/xrRender/Debug/dxPixEventWrapper.h"
 #include "Layers/xrRenderDX9/dx9HW.h"
