+++ src/Layers/xrRenderPC_R2/stdafx.h
@@ -10,7 +10,7 @@
 #include <d3d9.h>
 #include <d3dx9.h>
 
-#include "Layers/xrRender/xrD3DDefs.h"
+#include "Layers/xrRenderDX9/CommonTypes.h"
 #include "Layers/xrRender/Debug/dxPixEventWrapper.h"
 #include "Layers/xrRenderDX9/dx9HW.h"
 #include "Layers/xrRender/Shader.h"
